init seperation
