+++ components/PokemonCard.vue
@@ -1,39 +1,7 @@
 <script setup lang="ts">
-import { useI18n } from 'vue-i18n'
-import type moves from '../assets/pokemonMove.json'
-import type abilities from '../assets/pokemonAbility.json'
-import type items from '../assets/pokemonItem.json'
-
-type JSONKey<T extends Record<string, any>> = keyof T
-type JSONValue<T extends Record<string, any>> = T[JSONKey<T>]
-
-const { t } = useI18n()
+import { abilityProps, itemsProps, movesProps } from '~/composables/useAssetKeyToContext'
 
 const show = ref(false)
-
-function movesProps(move: JSONKey<typeof moves>, value: JSONValue<typeof moves>) {
-  const { type, basePower, category } = value
-  return {
-    title: `${t(`move.${move}`)}`,
-    subtitle: `${t(`type.${type ? type.toLowerCase() : type}`)}/${basePower}/${t(`${category}`)}`
-  }
-}
-
-function abilityProps(ability: JSONKey<typeof abilities>, value: JSONValue<typeof abilities>) {
-  const { effect } = value
-  return {
-    title: `${t(`ability.${ability}`)}`,
-    subtitle: effect
-  }
-}
-
-function itemsProps(item: JSONKey<typeof items>, value: JSONValue<typeof items>) {
-  const { effect } = value
-  return {
-    title: `${t(`item.${item}`)}`,
-    subtitle: effect
-  }
-}
 </script>
 
 <template>
