+++ package.json
@@ -40,6 +40,7 @@
     "nuxt": "^3.8.2",
     "sass": "^1.70.0",
     "sass-loader": "^14.1.0",
+    "typescript": "^5.4.2",
     "vite-plugin-vuetify": "^2.0.1",
     "vitest": "^1.2.1",
     "vue": "^3.3.12",
